+++ teensy/teensy-speaker.h
@@ -18,9 +18,9 @@ class TeensySpeaker : public PhysicalSpeaker {
 
   virtual void begin();
 
-  virtual void toggle(uint32_t c);
+  virtual void toggle(int64_t c);
   virtual void maintainSpeaker();
-  virtual void maintainSpeaker(uint32_t c, uint64_t microseconds);
+  virtual void maintainSpeaker(int64_t c, uint64_t microseconds);
 
   virtual void beginMixing();
   virtual void mixOutput(uint8_t v);
